move launcher errors under launcher package
